[output-mapping] FilesHelper::getFile moved to SliceHelper
